update: refactor package type structure
