+++ packages/rpc/src/index.ts
@@ -1,3 +1,3 @@
 export * from './rpc'
-export * from './utils'
+export * from './utils/misc'
 export * from './ws'
