+++ packages/rpc/src/utils/misc.ts
@@ -1,0 +1,3 @@
+export const unresolvablePromise = new Promise<void>(() => {})
+
+export const schedule = setTimeout
